ai-startup-evaluator/voice-interview-agent/app/agent.py
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

import datetime
import os
import logging
from typing import Dict, List, Optional, Any

import google.auth
from google.adk.agents import Agent

_, project_id = google.auth.default()
os.environ.setdefault("GOOGLE_CLOUD_PROJECT", project_id)
os.environ.setdefault("GOOGLE_CLOUD_LOCATION", "us-central1")
os.environ.setdefault("GOOGLE_GENAI_USE_VERTEXAI", "True")

logger = logging.getLogger(__name__)

def start_voice_session(founder_name: str) -> str:
    """Start a voice interview session"""
    return f"Started voice session for {founder_name}"

def process_voice_input(audio_data: str) -> str:
    """Process voice input from founder"""
    return f"Processed voice input: {audio_data[:50]}..."

def generate_voice_response(response: str) -> str:
    """Generate voice response to founder"""
    return f"Generated voice response: {response[:50]}..."

# Create the agent
root_agent = Agent(
    name="voice_interview_agent",
    model="gemini-2.5-flash",
    instruction="Voice Interview Agent - Advanced voice interview agent with real-time processing (OPTIONAL)",
    tools=[start_voice_session, process_voice_input, generate_voice_response],
)
